+++ tests/testthat/test-appMetadata.R
@@ -65,30 +65,43 @@ test_that("checkLayout() errors if primary doc & app.R", {
     "myscript.R" = ""
   ))
 
-  expect_snapshot(checkAppLayout(dir, "myscript.R"), error = TRUE)
-})
-
-test_that("checkLayout fails if no known structure", {
-  dir <- local_temp_app(list(
-    "data.txt" = "",
-    "cats.csv" = ""
-  ))
-
-  expect_snapshot(checkAppLayout(dir), error = TRUE)
+  expect_snapshot(checkAppLayout(dir, appPrimaryDoc = "myscript.R"), error = TRUE)
 })
 
 test_that("checkLayout succeeds with some common app structures", {
   rmd <- local_temp_app(list("foo.Rmd" = ""))
   expect_no_error(checkAppLayout(rmd))
 
+  qmd <- local_temp_app(list("foo.qmd" = ""))
+  expect_no_error(checkAppLayout(qmd))
+
+  # perhaps script.R is a pre-render script that generates *.qmd
+  project <- local_temp_app(list("_quarto.yml" = "", "script.R" = ""))
+  expect_no_error(checkAppLayout(project))
+
+  md <- local_temp_app(list("_quarto.yml" = "", "index.md" = ""))
+  expect_no_error(checkAppLayout(md))
+
   shiny1 <- local_temp_app(list("app.R" = ""))
-  expect_no_error(checkAppLayout(rmd))
+  expect_no_error(checkAppLayout(shiny1))
 
   shiny2 <- local_temp_app(list("server.R" = "", "ui.R" = ""))
-  expect_no_error(checkAppLayout(rmd))
+  expect_no_error(checkAppLayout(shiny2))
+
+  api <- local_temp_app(list("plumber.R" = ""))
+  expect_no_error(checkAppLayout(api))
 
   static <- local_temp_app(list("foo.html" = ""))
-  expect_no_error(checkAppLayout(rmd))
+  expect_no_error(checkAppLayout(static))
+
+  staticxml <- local_temp_app(list("data.xml" = ""))
+  expect_no_error(checkAppLayout(staticxml))
+
+  staticdata <- local_temp_app(list(
+    "data.txt" = "",
+    "cats.csv" = ""
+  ))
+  expect_no_error(checkAppLayout(staticdata))
 })
 
 # inferAppMode ------------------------------------------------------------