+++ NEWS.md
@@ -5,6 +5,11 @@
 * Do not use `getOption("available_packages_filters")` option when calling
   `available.packages()`. (#1002)
 
+* Improve reporting of errors returned by shinyapps.io. (#997)
+
+* Remove most directory layout validation checks. (#998)
+
+
 # rsconnect 1.1.0
 
 * Fixed analysis of directories that were smaller than the